go/keymanager: Move master and ephemeral secrets code to extension
