+++ go/consensus/cometbft/apps/keymanager/secrets/state/state_test.go
@@ -8,7 +8,7 @@ import (
 	beacon "github.com/oasisprotocol/oasis-core/go/beacon/api"
 	"github.com/oasisprotocol/oasis-core/go/common"
 	abciAPI "github.com/oasisprotocol/oasis-core/go/consensus/cometbft/api"
-	"github.com/oasisprotocol/oasis-core/go/keymanager/api"
+	"github.com/oasisprotocol/oasis-core/go/keymanager/secrets"
 )
 
 func TestMasterSecret(t *testing.T) {
@@ -25,19 +25,19 @@ func TestMasterSecret(t *testing.T) {
 		common.NewTestNamespaceFromSeed([]byte("runtime 1"), common.NamespaceKeyManager),
 		common.NewTestNamespaceFromSeed([]byte("runtime 2"), common.NamespaceKeyManager),
 	}
-	secrets := make([]*api.SignedEncryptedMasterSecret, 0, 10)
-	for i := 0; i < cap(secrets); i++ {
-		secret := api.SignedEncryptedMasterSecret{
-			Secret: api.EncryptedMasterSecret{
+	masterSecrets := make([]*secrets.SignedEncryptedMasterSecret, 0, 10)
+	for i := 0; i < cap(masterSecrets); i++ {
+		secret := secrets.SignedEncryptedMasterSecret{
+			Secret: secrets.EncryptedMasterSecret{
 				ID:         runtimes[i%2],
 				Generation: uint64(i),
 			},
 		}
-		secrets = append(secrets, &secret)
+		masterSecrets = append(masterSecrets, &secret)
 	}
 
 	// Test adding secrets.
-	for _, secret := range secrets {
+	for _, secret := range masterSecrets {
 		err := s.SetMasterSecret(ctx, secret)
 		require.NoError(err, "SetMasterSecret()")
 	}
@@ -46,10 +46,10 @@ func TestMasterSecret(t *testing.T) {
 	for i, runtime := range runtimes {
 		secret, err := s.MasterSecret(ctx, runtime)
 		require.NoError(err, "MasterSecret()")
-		require.Equal(secrets[8+i], secret, "last master secret should be kept")
+		require.Equal(masterSecrets[8+i], secret, "last master secret should be kept")
 	}
 	_, err := s.MasterSecret(ctx, common.Namespace{1, 2, 3})
-	require.EqualError(err, api.ErrNoSuchMasterSecret.Error(), "MasterSecret should error for non-existing secrets")
+	require.EqualError(err, secrets.ErrNoSuchMasterSecret.Error(), "MasterSecret should error for non-existing secrets")
 }
 
 func TestEphemeralSecret(t *testing.T) {
@@ -66,19 +66,19 @@ func TestEphemeralSecret(t *testing.T) {
 		common.NewTestNamespaceFromSeed([]byte("runtime 1"), common.NamespaceKeyManager),
 		common.NewTestNamespaceFromSeed([]byte("runtime 2"), common.NamespaceKeyManager),
 	}
-	secrets := make([]*api.SignedEncryptedEphemeralSecret, 0, 10)
-	for i := 0; i < cap(secrets); i++ {
-		secret := api.SignedEncryptedEphemeralSecret{
-			Secret: api.EncryptedEphemeralSecret{
+	masterSecrets := make([]*secrets.SignedEncryptedEphemeralSecret, 0, 10)
+	for i := 0; i < cap(masterSecrets); i++ {
+		secret := secrets.SignedEncryptedEphemeralSecret{
+			Secret: secrets.EncryptedEphemeralSecret{
 				ID:    runtimes[i%2],
 				Epoch: beacon.EpochTime(i),
 			},
 		}
-		secrets = append(secrets, &secret)
+		masterSecrets = append(masterSecrets, &secret)
 	}
 
 	// Test adding secrets.
-	for _, secret := range secrets {
+	for _, secret := range masterSecrets {
 		err := s.SetEphemeralSecret(ctx, secret)
 		require.NoError(err, "SetEphemeralSecret()")
 	}
@@ -87,8 +87,8 @@ func TestEphemeralSecret(t *testing.T) {
 	for i, runtime := range runtimes {
 		secret, err := s.EphemeralSecret(ctx, runtime)
 		require.NoError(err, "EphemeralSecret()")
-		require.Equal(secrets[8+i], secret, "last ephemeral secret should be kept")
+		require.Equal(masterSecrets[8+i], secret, "last ephemeral secret should be kept")
 	}
 	_, err := s.EphemeralSecret(ctx, common.Namespace{1, 2, 3})
-	require.EqualError(err, api.ErrNoSuchEphemeralSecret.Error(), "EphemeralSecret should error for non-existing secrets")
+	require.EqualError(err, secrets.ErrNoSuchEphemeralSecret.Error(), "EphemeralSecret should error for non-existing secrets")
 }
