+++ go/consensus/cometbft/apps/keymanager/secrets/status_test.go
@@ -1,4 +1,4 @@
-package keymanager
+package secrets
 
 import (
 	"testing"
@@ -15,6 +15,7 @@ import (
 	"github.com/oasisprotocol/oasis-core/go/common/version"
 	abciAPI "github.com/oasisprotocol/oasis-core/go/consensus/cometbft/api"
 	"github.com/oasisprotocol/oasis-core/go/keymanager/api"
+	"github.com/oasisprotocol/oasis-core/go/keymanager/secrets"
 	registry "github.com/oasisprotocol/oasis-core/go/registry/api"
 )
 
@@ -24,15 +25,10 @@ func TestGenerateStatus(t *testing.T) {
 	ctx := appState.NewContext(abciAPI.ContextEndBlock)
 	defer ctx.Close()
 
-	// Prepare app.
-	app := &keymanagerApplication{
-		state: appState,
-	}
-
 	// Prepare vars.
 	params := &registry.ConsensusParameters{}
-	policy := api.SignedPolicySGX{
-		Policy: api.PolicySGX{
+	policy := secrets.SignedPolicySGX{
+		Policy: secrets.PolicySGX{
 			Serial: 1,
 		},
 	}
@@ -42,20 +38,20 @@ func TestGenerateStatus(t *testing.T) {
 
 	// Prepare two responses so that we can test nodes running different versions.
 	rakSigner := api.TestSigners[0]
-	initResponse := api.InitResponse{
+	initResponse := secrets.InitResponse{
 		IsSecure:       true,
 		Checksum:       checksum,
 		PolicyChecksum: policyChecksum[:],
 	}
-	sigInitResponse, err := api.SignInitResponse(rakSigner, &initResponse)
+	sigInitResponse, err := secrets.SignInitResponse(rakSigner, &initResponse)
 	require.NoError(t, err, "SignInitResponse")
 
 	initResponse.Checksum = nil
-	sigInitResponseSecure, err := api.SignInitResponse(rakSigner, &initResponse)
+	sigInitResponseSecure, err := secrets.SignInitResponse(rakSigner, &initResponse)
 	require.NoError(t, err, "SignInitResponse")
 
 	initResponse.IsSecure = false
-	sigInitResponseInsecure, err := api.SignInitResponse(rakSigner, &initResponse)
+	sigInitResponseInsecure, err := secrets.SignInitResponse(rakSigner, &initResponse)
 	require.NoError(t, err, "SignInitResponse")
 
 	// Two key manager runtimes, one compute runtime.
@@ -65,14 +61,14 @@ func TestGenerateStatus(t *testing.T) {
 	require.NoError(t, runtimeIDs[2].UnmarshalHex("8000000000000000000000000000000000000000000000000000000000000002"), "runtime 2")
 
 	// Initial key manager statuses.
-	initializedStatus := &api.Status{
+	initializedStatus := &secrets.Status{
 		ID:            runtimeIDs[0],
 		IsInitialized: true,
 		IsSecure:      true,
 		Checksum:      checksum,
 		Policy:        &policy,
 	}
-	uninitializedStatus := &api.Status{
+	uninitializedStatus := &secrets.Status{
 		ID:     runtimeIDs[0],
 		Policy: &policy,
 	}
@@ -211,10 +207,10 @@ func TestGenerateStatus(t *testing.T) {
 	t.Run("No nodes", func(t *testing.T) {
 		require := require.New(t)
 
-		newStatus := app.generateStatus(ctx, runtimes[0], uninitializedStatus, nil, nodes[0:6], params, epoch)
+		newStatus := generateStatus(ctx, runtimes[0], uninitializedStatus, nil, nodes[0:6], params, epoch)
 		require.Equal(uninitializedStatus, newStatus, "key manager committee should be empty")
 
-		newStatus = app.generateStatus(ctx, runtimes[0], initializedStatus, nil, nodes[0:6], params, epoch)
+		newStatus = generateStatus(ctx, runtimes[0], initializedStatus, nil, nodes[0:6], params, epoch)
 		require.Equal(initializedStatus, newStatus, "key manager committee should be empty")
 	})
 
@@ -222,23 +218,23 @@ func TestGenerateStatus(t *testing.T) {
 		require := require.New(t)
 
 		// Node 6 (secure = false)
-		expStatus := &api.Status{
+		expStatus := &secrets.Status{
 			ID:            runtimeIDs[0],
 			IsInitialized: true,
 			IsSecure:      false,
 			Policy:        &policy,
 			Nodes:         []signature.PublicKey{nodes[6].ID},
 		}
-		newStatus := app.generateStatus(ctx, runtimes[0], uninitializedStatus, nil, nodes[6:7], params, epoch)
+		newStatus := generateStatus(ctx, runtimes[0], uninitializedStatus, nil, nodes[6:7], params, epoch)
 		require.Equal(expStatus, newStatus, "node 6 should form the committee if key manager not initialized")
 
-		newStatus = app.generateStatus(ctx, runtimes[0], expStatus, nil, nodes[6:7], params, epoch)
+		newStatus = generateStatus(ctx, runtimes[0], expStatus, nil, nodes[6:7], params, epoch)
 		require.Equal(expStatus, newStatus, "node 6 should form the committee if key manager is not secure")
 
 		expStatus.IsSecure = true
 		expStatus.Checksum = checksum
 		expStatus.Nodes = nil
-		newStatus = app.generateStatus(ctx, runtimes[0], initializedStatus, nil, nodes[6:7], params, epoch)
+		newStatus = generateStatus(ctx, runtimes[0], initializedStatus, nil, nodes[6:7], params, epoch)
 		require.Equal(expStatus, newStatus, "node 6 should not be added to the committee if key manager is secure or checksum differs")
 	})
 
@@ -248,31 +244,31 @@ func TestGenerateStatus(t *testing.T) {
 		// The first node is the source of truth when constructing a committee.
 		// If the node 6 is processed before node 7, the latter won't be accepted as it is secure.
 		// Nodes 8 and 9 cannot be a part of the committee as their checksum differs.
-		expStatus := &api.Status{
+		expStatus := &secrets.Status{
 			ID:            runtimeIDs[0],
 			IsInitialized: true,
 			IsSecure:      false,
 			Policy:        &policy,
 			Nodes:         []signature.PublicKey{nodes[6].ID},
 		}
-		newStatus := app.generateStatus(ctx, runtimes[0], uninitializedStatus, nil, nodes, params, epoch)
+		newStatus := generateStatus(ctx, runtimes[0], uninitializedStatus, nil, nodes, params, epoch)
 		require.Equal(expStatus, newStatus, "node 6 should be the source of truth and form the committee")
 
 		// If the order is reversed, it should be the other way around.
 		expStatus.IsSecure = true
 		expStatus.Nodes = []signature.PublicKey{nodes[7].ID}
-		newStatus = app.generateStatus(ctx, runtimes[0], uninitializedStatus, nil, reverse(nodes), params, epoch)
+		newStatus = generateStatus(ctx, runtimes[0], uninitializedStatus, nil, reverse(nodes), params, epoch)
 		require.Equal(expStatus, newStatus, "node 7 should be the source of truth and form the committee")
 
 		// If the key manager is already initialized as secure with a checksum, then all nodes
 		// except 8 and 9 are ignored.
 		expStatus.Checksum = checksum
 		expStatus.Nodes = []signature.PublicKey{nodes[8].ID, nodes[9].ID}
-		newStatus = app.generateStatus(ctx, runtimes[0], initializedStatus, nil, nodes, params, epoch)
+		newStatus = generateStatus(ctx, runtimes[0], initializedStatus, nil, nodes, params, epoch)
 		require.Equal(expStatus, newStatus, "node 7 and 8 should form the committee if key manager is initialized as secure")
 
 		// The second key manager.
-		expStatus = &api.Status{
+		expStatus = &secrets.Status{
 			ID:            runtimeIDs[1],
 			IsInitialized: true,
 			IsSecure:      true,
@@ -281,7 +277,7 @@ func TestGenerateStatus(t *testing.T) {
 			Nodes:         []signature.PublicKey{nodes[4].ID, nodes[9].ID},
 		}
 		initializedStatus.ID = runtimeIDs[1]
-		newStatus = app.generateStatus(ctx, runtimes[1], initializedStatus, nil, nodes, params, epoch)
+		newStatus = generateStatus(ctx, runtimes[1], initializedStatus, nil, nodes, params, epoch)
 		require.Equal(expStatus, newStatus, "node 4 and 9 should form the committee")
 	})
 }
