+++ go/keymanager/secrets/api_test.go
@@ -1,4 +1,4 @@
-package api
+package secrets
 
 import (
 	"testing"
