+++ go/keymanager/secrets/sanity_check.go
@@ -1,4 +1,4 @@
-package api
+package secrets
 
 import (
 	"fmt"
