+++ go/keymanager/secrets/secret.go
@@ -1,4 +1,4 @@
-package api
+package secrets
 
 import (
 	"fmt"
