+++ go/keymanager/secrets/secret_test.go
@@ -1,4 +1,4 @@
-package api
+package secrets
 
 import (
 	"crypto/sha512"
