+++ runtime/src/consensus/state/keymanager.rs
@@ -154,8 +154,8 @@ mod test {
     };
 
     #[test]
-    fn test_keymanager_state_interop() {
-        // Keep in sync with go/consensus/cometbft/apps/keymanager/state/interop/interop.go.
+    fn test_keymanager_secrets_state_interop() {
+        // Keep in sync with go/consensus/cometbft/apps/keymanager/secrets/state/interop/interop.go.
         // If mock consensus state changes, update the root hash bellow.
         // See protocol server stdout for hash.
         // To make the hash show up during tests, run "cargo test" as
